+++ src/js/stores/AppsStore.js
@@ -6,6 +6,7 @@ import appScheme from "../stores/schemes/appScheme";
 import AppTypes from "../constants/AppTypes";
 import ContainerConstants from "../constants/ContainerConstants";
 import AppStatus from "../constants/AppStatus";
+import groupScheme from "../stores/schemes/groupScheme";
 import HealthStatus from "../constants/HealthStatus";
 import LocalVolumesConstants from "../constants/LocalVolumesConstants";
 import TasksEvents from "../events/TasksEvents";
@@ -147,29 +148,53 @@ function processApp(app) {
   return app;
 }
 
-function processEmptyGroup(group) {
-  return Object.assign(processApp({id: group.id}), {
-    isGroup: true,
-    status: null
+function calculateTotalStats(parent, child) {
+  Object.keys(parent).forEach(key => {
+    if (key === "health") {
+      parent[key].forEach((value, index) => {
+        if (child[key] != null && child[key][index] != null) {
+          value.quantity += child[key][index].quantity;
+        }
+      });
+      return;
+    }
+    if (Util.isNumber(parent[key])) {
+      parent[key] += child[key];
+    }
   });
 }
 
-// TODO: refactor to adhere to /groups endpoint structures.
-// https://github.com/mesosphere/marathon/issues/3383
-function processAppsAndEmptyGroups(group, apps = []) {
-  if (group.apps != null && group.apps.length > 0) {
-    group.apps.forEach(app => {
-      apps.push(processApp(app));
-    });
-  } else if (group.id !== "/") {
-    apps.push(processEmptyGroup(group));
+// TODO: Refactor this method as a part of the issue
+// https://github.com/mesosphere/marathon#3565
+function generateFlatList(group, root = true) {
+  var items = [];
+  var GroupData = Object.assign(
+    {},
+    Util.deepCopy(groupScheme),
+    {id: group.id, isGroup: true}
+  );
+  if (group.groups != null) {
+    items = items.concat(group.groups.reduce((groups, group) => {
+      var childGroupData = generateFlatList(group, false);
+      calculateTotalStats(GroupData, childGroupData.groupData);
+      return groups.concat(childGroupData.items);
+    }, []));
   }
-
-  if (group.groups != null && group.groups.length > 0) {
-    group.groups.forEach(group => processAppsAndEmptyGroups(group, apps));
+  if (group.apps != null) {
+    items = items.concat(group.apps.map(app => {
+      app =  processApp(app);
+      calculateTotalStats(GroupData, app);
+      return app;
+    }));
   }
-
-  return apps;
+  if (root) {
+    return items;
+  }
+  items.push(GroupData);
+  return {
+    items: items,
+    groupData: GroupData
+  };
 }
 
 function applyAppDelayStatus(app, queue) {
@@ -295,7 +320,7 @@ QueueStore.on(QueueEvents.CHANGE, function () {
 AppDispatcher.register(function (action) {
   switch (action.actionType) {
     case AppsEvents.REQUEST_APPS:
-      storeData.apps = processAppsAndEmptyGroups(action.data.body);
+      storeData.apps = generateFlatList(action.data.body);
       applyAppDelayStatusOnAllApps(storeData.apps, QueueStore.queue);
       AppsStore.emit(AppsEvents.CHANGE);
       break;